+++ assets/javascripts/kitten/components/information/alert/stories.js
@@ -2,7 +2,6 @@ import React from 'react'
 import { Alert } from './'
 import { action } from '@storybook/addon-actions'
 import { DocsPage } from 'storybook/docs-page'
-import { CrossIconNext, COLORS } from 'kitten'
 
 export default {
   title: 'Information/Alert',
@@ -14,31 +13,19 @@ export default {
   },
   args: {
     show: true,
-    error: false,
-    success: false,
-    warning: false,
     closeButton: true,
     closeButtonLabel: 'Close Button label',
     onAfterClose: action('onAfterClose'),
     children: 'Modifications effectuées, lien ici, Modifications effectuées,',
-    icon: <CrossIconNext width={12} height={12} color={COLORS.background1} />,
+    icon: null,
     center: false,
+    iconHasBorder: true,
+    displayIcon: true,
+    status: 'info',
   },
   argTypes: {
     show: {
       name: 'show',
-      control: 'boolean',
-    },
-    error: {
-      name: 'error',
-      control: 'boolean',
-    },
-    success: {
-      name: 'success',
-      control: 'boolean',
-    },
-    warning: {
-      name: 'warning',
       control: 'boolean',
     },
     closeButton: {
@@ -59,6 +46,14 @@ export default {
     },
     center: {
       control: 'boolean',
+    },
+    displayIcon: { control: 'boolean' },
+    iconHasBorder: {
+      control: 'boolean',
+    },
+    status: {
+      control: 'select',
+      options: ['info', 'success', 'warning', 'danger'],
     },
   },
 }